Miscellaneous improvements

+++ nosh/dirutils.py
@@ -17,13 +17,19 @@ def current_directory(new_dir):
     os.chdir(cur_dir)
 
 @contextlib.contextmanager
-def temp_directory():
+def temp_directory(clean=True):
     '''Context manager returning a temporary directory name.  This
     directory, and all its contents, is deleted at the end of the
     context.
 
+    Parameters
+    ----------
+    clean : bool
+        Whether to delete the temp dir after use. Defaults to True.
+        The main reason to set False is for testing purposes.
     '''
     temp_dir = tempfile.mkdtemp()
     yield temp_dir
-    shutil.rmtree(temp_dir)
+    if clean:
+        shutil.rmtree(temp_dir)
